Make referring to the NAND parts consistent.
Original usage was a mixture of parts and chunks. As the joiner scripts use parts in their name, for consistency, this has
been changed.

Also noted that Hekate makes 1 or 2 GB parts. While the exact reasoning for when either is used is irrelevant to the
guide, it does not make 4GB parts.