Update Lockpick section (#58)
* Update Lockpick section 

Update the steps for dumping the keys with the latest release of Lockpick which includes choosing between sys and emummc

* Update Lockpick section

Improved the previous edit by adding that one should navigate through the options with the volume buttons. Also fixed the order of the steps which were not correct in the previous file.