Updated instructions on migration to different SD
For the file based emunand I've added the step with the fix of the archive bit
Some users may have issues with lost archive bits during the file-transfer. I believe it is OS dependent, I've encountered this issue on MacOS, so it also may occur for linux. If that happens - emunand boots alright all the applications are unable to start, performing the `Fix Archive Bit` in hekate helps with that issue.
Though this step is not mandatory for all cases (should be ok with windows) - I've added it as a last step since it'll do no harm even when not necessary, but on the other side the following sequence of action should leave end-user with a working system no matter what was used for the file-transfer.