treewide: use `1.` numbering for all ordered lists
Following specifications at [0] the formatting for lists "have no
effect on the HTML output Markdown produces".

This means we can consolidate all ordered lists to `1.`, which makes it
easier for the end user to add and remove steps and retain the
numbering.

It should be noted some Markdown implementations may use the first
number as a starting number for ordered lists, so it's best to start
with `1.`.

[0]: https://daringfireball.net/projects/markdown/syntax#list